fix not checking for other dimension when copying
This is a great example of an optimization that went way wrong and
probably also is an example of weak argument checking. Imagine two
ropes like this:

upper : [[0; 0; 0]], [[0; 0; 0; 0]], Empty, Empty
lower : [[0; 0; 0; 0]], [[0; 0; 0]], Empty, Empty

Here, vcat would originally have used cat1 to merge the arrays, But cat2
choses the min of the width of both arrays! Therefore, the result would
wrongfully be:

[[0; 0; 0];
 [0; 0; 0]], [[0; 0; 0];
              [0; 0; 0]], Empty, Empty

The correct measure would have been to avoid copying. This commit fixes
this problem.