implement parallel wrapper
This is mostly a mechanical change; the implementation is taken from the
sequential wrapper and all constructors have been changed by
replace-string. Then, the functions which can be executed in parallel
were replaced by their parallel counterparts. All methods still
sequential, either due to the nature of the algorithm or due to missing
implementations, were marked with a comment.