remove fold variants
Fold is problematic in many ways. First, is not parallelizable (except
for line- or row-wise parallelization, which is boring). Second, it is
quite difficult to implement efficiently. It requires repeated
materialization of intermediate values and it gets even more
complicated in the presence of sparseness.

If we at some point want to reintroduce fold variants, it is enough to
revert this commit.