fix materialize sometimes retaining se branch
This can happen if materializing ne and sw both evaluate to
empty. For instance, on `slice 0 0 1 1 qr`, the retained h and w
values for the se case will we 1 and 1, because ne and sw have 0 rows
and 0 cols, respectively. In the old expression, we would then call
`materialize -r -c 1 1` for some r and c, which would result in a
slice of se of size 1 times 1.

Using conditionals is slower, but it is correct. It may be possible to
replace those by bit-operations.